Allow new genres to be added. (Issue #4)
This is a fix for issue #4 that allows the user to create new genres for
movies and TV shows.

One small note, the VideoLibrary.GetGenres rpc call only returns a list
of genres that contain one or more movies/shows. So don't be suprised if
some of the standard genres are missing.