Add repository to addon.xml for auto updates
I've setup a repository on my server for now. It'll check for new version tags (vX.Y.Z) every 6 hours and update the repository as appropriate. I've set it to 6 hours so I don't hit the GitHub servers to often and because Kodi only checks for repository updates every 24 hours.

With this change users will only have to manually install the add-on once. Once it's installed it'll check for updates automatically.